moved egress_proxy settings
